# project-backend/swagger.yaml
swagger: "2.0"
info:
  title: Toohak API
  description: |
    API description in for Toohak.
    All error strings are required to be clear descriptive message. The descriptive message won't change the marks you receive when being automarked as long as you make an attempt for each one to be custom, but they will improve the usability of the frontend.
    See the schema tab for greater explanations on each attribute of an object.
  version: 1.0.0
schemes:
  - http
...
components:
...
  header:
    Session:
      in: header
      name: session
      required: true
      schema:
        type: string
...
  body:
...
    Register:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          email:
            $ref: '#/components/prim/Email'
          password:
            $ref: '#/components/prim/Password'
          nameFirst:
            $ref: '#/components/prim/NameFirst'
          nameLast:
            $ref: '#/components/prim/NameLast'
    Login:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          email:
            $ref: '#/components/prim/Email'
          password:
            $ref: '#/components/prim/Password'
...
  return:
    Session:
      type: object
      properties:
        session:
          $ref: '#/components/prim/Session'
    Error:
      properties:
        error:
          type: string
            - error
          $ref: '#/components/prim/ErrorType'
        message:
          type: string
            - message
          $ref: '#/components/prim/ErrorMessage'
    Empty:
      type: object
    User:
      type: object
      properties:
        user:
          $ref: '#/components/group/User'
    QuizList:
      type: object
      properties:
        quizzes:
          $ref: '#/components/group/SmallQuizzes'
...
paths:
  /v1/admin/auth/register:
    post:
      summary: Register a new admin user
      description: Takes in information about a new admin user and registers them in the system. This route is not relevant to guests who want to play a particular quiz, but is used for the creation of accounts of people who manage quizzes.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/body/Register'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Session'
        400:
          description: |
            If any of the following are true:
            - **INVALID_EMAIL**
              - Email address is used by another user
              - Email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - **INVALID_FIRST_NAME**
              - NameFirst contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
              - NameFirst is less than 2 characters or more than 20 characters
            - **INVALID_LAST_NAME**
              - NameLast contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
              - NameLast is less than 2 characters or more than 20 characters
            - **INVALID_PASSWORD**
              - Password is less than 8 characters
              - Password does not contain at least one number and at least one letter
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/auth/login:
    post:
      summary: Login an admin user
      description: Takes in information about an admin user to determine if they can log in to manage quizzes. This route is not relevant to guests who want to play a particular quiz, but is used for the creation of accounts of people who manage quizzes. 
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/body/Login'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Session'
        400:
          description: |
            If any of the following are true:
            - **INVALID_CREDENTIALS**
              - Email address does not exist
              - Password is not correct for the given email
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/auth/logout:
    post:
      summary: Logs out an admin user who has an active user session.
      description: Should be called with a session that is returned after either a login or register has been made.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (New)"
      parameters:
        - $ref: '#/components/header/Session'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/user/details:
    get:
      summary: Get the details of an admin user.
      description: For the given admin user that is logged in, return all of the relevant details.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/Session'
      responses:
        200:
          description: |
            - "name" is the first and last name concatenated with a single space between them
            - numSuccessfulLogins includes logins direct via registration, and is counted from the moment of registration starting at 1
            - numFailedPasswordsSinceLastLogin is reset every time they have a successful login, and simply counts the number of attempted logins that failed due to incorrect password, only since the last login
          schema:
            $ref: '#/components/return/User'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
    put:
      summary: Update the details of an admin user.
      description: Given a set of properties, update those properties of this logged in admin user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/Session'
        - $ref: '#/components/body/UserDetailsPutIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - **INVALID_EMAIL**
              - Email is currently used by another user (excluding the current authorised user)
              - Email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - **INVALID_FIRST_NAME**
              - NameFirst contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
              - NameFirst is less than 2 characters or more than 20 characters
            - **INVALID_LAST_NAME**
              - NameLast contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
              - NameLast is less than 2 characters or more than 20 characters
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/user/password:
    put:
      summary: Update the password of this admin user.
      description: Given details relating to a password change, update the password of a logged in user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/Session'
        - $ref: '#/components/body/UserPasswordIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - **INVALID_OLD_PASSWORD**
              - Old Password is not the correct old password
            - **INVALID_NEW_PASSWORD**
              - Old Password and New Password match exactly
              - New Password has already been used before by this user
              - New Password is less than 8 characters
              - New Password does not contain at least one number and at least one letter
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/quiz/list:
    get:
      summary: Lists all user's quizzes
      description: Provide a list of all quizzes that are owned by the currently logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/Session'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizList'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/quiz:
    post:
      summary: Create a new quiz
      description: Given basic details about a new quiz, create one for the logged in user
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/Session'
        - $ref: '#/components/body/QuizNewIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizNew'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - **INVALID_QUIZ_NAME**
              - Name contains invalid characters. Valid characters are alphanumeric and spaces
              - Name is either less than 3 characters long or more than 30 characters long
            - **DUPLICATE_QUIZ_NAME**
              - Name is already used by the current logged in user for another quiz
            - **INVALID_DESCRIPTION**
              - Description is more than 100 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/quiz/{quizid}:
    delete:
      summary: Delete a quiz
      description: |
       - Given a particular quiz, delete a quiz
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/header/Session'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: |
            - **INVALID_QUIZ_ID**
              - Valid session is provided, but user is not an owner of this quiz, or quiz doesn't exist
          schema:
            $ref: '#/components/return/Error'
    get:
      summary: Get info about current quiz
      description: Get all of the relevant information about the current quiz including questions
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/header/Session'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/QuizSingleIter3'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: |
            - **INVALID_QUIZ_ID**
              - Valid session is provided, but user is not an owner of this quiz, or quiz doesn't exist
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/quiz/{quizid}/name:
    put:
      summary: Update quiz name
      description: Update the name of the relevant quiz
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/header/Session'
        - $ref: '#/components/body/QuizUpdateNameIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: |
            - **INVALID_QUIZ_ID**
              - Valid session is provided, but user is not an owner of this quiz, or quiz doesn't exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - **INVALID_QUIZ_NAME**
              - Name contains invalid characters. Valid characters are alphanumeric and spaces
              - Name is either less than 3 characters long or more than 30 characters long
            - **DUPLICATE_QUIZ_NAME**
              - Name is already used by the current logged in user for another quiz
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/quiz/{quizid}/description:
    put:
      summary: Update quiz description
      description: Update the description of the relevant quiz
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/Session'
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/body/QuizUpdateDescriptionIter3'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: |
            - **INVALID_QUIZ_ID**
              - Valid session is provided, but user is not an owner of this quiz, or quiz doesn't exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - **INVALID_DESCRIPTION**
              - Description is more than 100 characters in length (note: empty strings are OK)
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/quiz/{quizid}/thumbnail:
    put:
      summary: Update the quiz thumbnail
      description: |
        Update the thumbnail for the quiz.
        When this route is called, the timeLastEdited is updated.
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 2 (New)"
      parameters:
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/header/Session'
        - $ref: '#/components/body/QuizThumbnail'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: |
            - **INVALID_QUIZ_ID**
              - Valid session is provided, but user is not an owner of this quiz, or quiz doesn't exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - **INVALID_THUMBNAIL**
              - The thumbnailUrl does not end with one of the following filetypes (case insensitive): jpg, jpeg, png
              - The thumbnailUrl does not begin with 'http://' or 'https://'
          schema:
            $ref: '#/components/return/Error'
...
  /v2/admin/quiz/{quizid}/question/{questionid}:
    delete:
      summary: Delete quiz question
      description: Delete a particular question from a quiz
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/path/QuestionId'
        - $ref: '#/components/header/Session'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: |
            - **INVALID_QUIZ_ID**
              - Valid session is provided, but user is not an owner of this quiz, or quiz doesn't exist
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - **INVALID_QUESTION_ID**
              - Question Id does not refer to a valid question within this quiz
            - **ACTIVE_GAME_EXISTS**
              - Any game for this quiz is not in END state
          schema:
            $ref: '#/components/return/Error'
      security:
        - Iteration 3 Session Authentication: []
...
  /v2/admin/quiz/{quizid}:
    delete:
      summary: Delete a quiz
      description: Given a particular quiz, delete a quiz
      tags:
        # - "By Topic: Quiz Management"
        - "Iteration 3 (Modified)"
      parameters:
        - $ref: '#/components/path/QuizId'
        - $ref: '#/components/header/Session'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        400:
          description: |
            - **ACTIVE_GAME_EXISTS**
              - Any game for this quiz is not in END state
          schema:
            $ref: '#/components/return/Error'
        401:
          description: |
            - **UNAUTHORISED**
              - Session is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: |
            - **INVALID_QUIZ_ID**
              - Valid session is provided, but user is not an owner of this quiz, or quiz doesn't exist
          schema:
            $ref: '#/components/return/Error'
...